Re #5056. VC dosn't think that size_t is unsigned long

+++ Code/Mantid/Framework/DataObjects/inc/MantidDataObjects/SplittersWorkspace.h
@@ -53,7 +53,7 @@ namespace DataObjects
 
     size_t getNumberSplitters();
 
-    bool removeSplitter(unsigned long);
+    bool removeSplitter(size_t);
 
   };
